Extract the logic into two different components and expose a a new image list interface without the list context
