+++ react/ImageList.tsx
@@ -1,25 +1,16 @@
-import { useDevice } from 'vtex.device-detector'
-import { useListContext, ListContextProvider } from 'vtex.list-context'
-
 import React from 'react'
-import { defineMessages } from 'react-intl'
+import Image from "./Image"
+import { useDevice } from "vtex.device-detector"
+import { IMAGE_LIST_SCHEMA } from "./utils/schema"
 
-import Image from './Image'
-
-interface Props {
+interface ImageListProps {
   images: Image[]
   height: number
-  listWithoutContext: boolean
 }
 
-const ImageList: StorefrontFunctionComponent<Props> = ({
-  images,
-  height = 420,
-  children,
-  listWithoutContext = false,
-}) => {
+const ImageList = (props: ImageListProps) => {
+  const { images, height = 420 } = props
   const { isMobile } = useDevice()
-  const list = useListContext()?.list ?? []
 
   const imageListContent = images.map(
     (
@@ -47,78 +38,13 @@ const ImageList: StorefrontFunctionComponent<Props> = ({
     )
   )
 
-  const newListContextValue = list.concat(imageListContent)
-
-  return listWithoutContext ? (
-    <>{newListContextValue}</>
-  ) : (
-    <ListContextProvider list={newListContextValue}>
-      {children}
-    </ListContextProvider>
+  return (
+    <React.Fragment>
+      {imageListContent}
+    </React.Fragment>
   )
 }
 
-const messages = defineMessages({
-  title: { id: 'admin/editor.image-list.title', defaultMessage: '' },
-  description: {
-    id: 'admin/editor.image-list.description',
-    defaultMessage: '',
-  },
-  imagesImageTitle: {
-    id: 'admin/editor.image-list.images.image.title',
-    defaultMessage: '',
-  },
-  imagesMobileImageTitle: {
-    id: 'admin/editor.image-list.images.mobileImage.title',
-    defaultMessage: '',
-  },
-  imagesImageDescription: {
-    id: 'admin/editor.image-list.images.description.title',
-    defaultMessage: '',
-  },
-  imagesImageAttributeTitle: {
-    id: 'admin/editor.image-list.images.title.title',
-    defaultMessage: '',
-  },
-  imagesImageLinkUrl: {
-    id: 'admin/editor.image-list.images.link.url.title',
-    defaultMessage: '',
-  },
-  imagesImageLinkOpenNewTab: {
-    id: 'admin/editor.image-list.images.link.openNewTab.title',
-    defaultMessage: '',
-  },
-  imagesImageLinkNoFollow: {
-    id: 'admin/editor.image-list.images.link.noFollow.title',
-    defaultMessage: '',
-  },
-  imagesImageLinkTitle: {
-    id: 'admin/editor.image-list.images.link.title.title',
-    defaultMessage: '',
-  },
-  imagesTitle: {
-    id: 'admin/editor.image-list.images.title',
-    defaultMessage: '',
-  },
-  heightTitle: {
-    id: 'admin/editor.image-list.height.title',
-    defaultMessage: '',
-  },
-})
-
-ImageList.schema = {
-  title: messages.title.id,
-  description: messages.description.id,
-  type: 'object',
-  properties: {
-    height: {
-      default: 420,
-      enum: [420, 440],
-      isLayout: true,
-      title: messages.heightTitle.id,
-      type: 'number',
-    },
-  },
-}
+ImageList.schema = IMAGE_LIST_SCHEMA
 
 export default ImageList
